migrated from jogl to skiko

+++ system/src/main/kotlin/common/Bounds.kt
@@ -1,10 +1,10 @@
 package common
 
 data class Bounds(
-    val x1: Double,
-    val y1: Double,
-    val x2: Double,
-    val y2: Double,
+    val x1: Float,
+    val y1: Float,
+    val x2: Float,
+    val y2: Float,
     val onClick: (() -> Unit)?,
     val onHold: (() -> Unit)?
 )
